Made Projectile class file seperate
